requirements.txt
django-environ==0.4.4
django-heroku==0.3.1
django-imagekit==4.0.2
Django==2.0.3
gunicorn==19.7.1
line-bot-sdk==1.5.0
Pillow==5.1.0
